django-google-bard-ai/src/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.5
django-autoslug==1.9.8
django-razorpay==1.1.6
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.4
frozenlist==1.4.0
future==0.18.2
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.12.0 
google-crc32c==1.5.0
google-generativeai==0.2.2
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
openai==0.28.1
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.4.0
platformdirs==3.11.0
proto-plus==1.22.3
protobuf==4.24.4
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
razorpay==1.4.1
redis==5.0.1
referencing==0.30.2
reportlab==3.6.8
requests==2.25.1
rpds-py==0.10.6
rsa==4.9
SecretStorage==3.3.1
shapely==2.0.2
six==1.16.0
sqlparse==0.4.3
systemd-python==234
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2022.7
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
